// src/app/service/models/simulation-request-global.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SimulationDtoRequest } from '../models/simulation-dto-request';
import { SimulationDtoResponse } from '../models/simulation-dto-response';
export interface SimulationRequestGlobal {
  request?: SimulationDtoRequest;
  response?: SimulationDtoResponse;
}